mig_app (mig v0.1.0)

This module implements the mig application.

Functions

start(StartType, StartArgs)
(since OTP 27.0)

-spec start(StartType :: application:start_type(), StartArgs :: term()) ->
               {ok, pid()} | ignore | {error, _}.

-callback start(StartType :: start_type(), StartArgs :: term()) ->
    {ok, pid()} | {ok, pid(), State :: term()} | {error, Reason :: term()}.

This function is called whenever an application is started using start/1,2, and is to start the processes of the application. If the application is structured according to the OTP design principles as a supervision tree, this means starting the top supervisor of the tree.

stop(State)
(since OTP 27.0)

-spec stop(State :: term()) -> term().

-callback stop(State :: term()) -> term().

This function is called whenever an application has stopped. It is intended to be the opposite of Module:start/2 and is to do any necessary cleaning up. The return value is ignored.